Fix failing tests
The GitHub Actions CI has been failing for a while because the output is
not rendering colored. This is because `test.js` was setting
`FORCE_COLOR=1` but this happens after the `import` of `chalk` and thus
also after the import within chalk of
[`supports-color`](https://www.npmjs.com/package/supports-color)

JS likes imports to be at the top of files before any other logic, so
it's not really possible to set `FORCE_COLOR` any earlier within
`test.js`. But it *is* possible to set it in `package.json` while
invoking `ava`...

Fixes: GH-27